% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group-age.R
\name{group_age_standard}
\alias{group_age_standard}
\title{Create Standard Demographic Age Groups}
\usage{
group_age_standard(
  dt,
  var,
  type = "standard",
  new_var = NULL,
  copy = FALSE,
  missing_values = NULL,
  ...
)
}
\arguments{
\item{dt}{data.table. The input data.table object.}

\item{var}{Character string. Name of the age variable.}

\item{type}{Character string. Type of age grouping. Options are:
\itemize{
\item "standard": 0-17, 18-34, 35-54, 55-74, 75+
\item "young": 0-2, 3-5, 6-12, 13-17, 18+
\item "young30": 16-20, 21-25, 26-30, 31+
\item "young34": 16-20, 21-25, 26-30, 31-34, 35+
\item "older": <65, 65-74, 75-84, 85+
\item "working": <18, 18-64, 65+
\item "unodc": 16-17, 18-24, 25-34, 35-64, 65+
\item "rusund": 16-24, 25-34, 35-44, 45-54, 55-64, 65-79
}}

\item{new_var}{Character string, optional. Name for the new variable.}

\item{copy}{Logical. Return a copy instead of modifying by reference?}

\item{missing_values}{Numeric vector, optional. Values that should be treated
as missing (e.g., survey codes). These values will be converted to NA
before categorization. Default is NULL.}

\item{...}{Other arguments in \code{group_age()}}
}
\value{
Modified data.table with age group variable added.
}
\description{
A convenience wrapper around \code{group_age()} that creates commonly used
demographic age categories.
}
\examples{
library(data.table)

# Create sample data
dt <- data.table(
  id = 1:100,
  age = sample(18:80, 100, replace = TRUE),
  gender = sample(c("M", "F"), 100, replace = TRUE)
)

# Create standard demographic groups
dt_demo <- group_age_standard(dt, "age", type = "standard")

# Create working-age focused groups
dt_work <- group_age_standard(dt, "age", type = "working", copy = TRUE)

}
