% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case.R
\name{is_case}
\alias{is_case}
\title{Create a binary indicator from one or more source variables}
\usage{
is_case(dt, varfrom, varto, val = 1)
}
\arguments{
\item{dt}{A \code{data.table}.}

\item{varfrom}{A character vector containing one or more source column names.}

\item{varto}{A character string specifying the name of the output column.}

\item{val}{A value used to identify cases. Defaults to \code{1}.}
}
\value{
A copy of \code{dt} with an additional binary column specified by \code{varto}.
}
\description{
Creates a new binary variable based on whether one or more source
variables equal a specified value. Missing values are treated as
non-matches.
}
\details{
For a single source variable, the output is:
\itemize{
\item 1 if the variable equals \code{val}
\item 0 if the variable does not equal \code{val}
\item 0 if the variable is \code{NA}
}

For multiple source variables, the output is:
\itemize{
\item 1 if at least one variable equals \code{val}
\item 0 otherwise
}
}
\examples{
library(data.table)

dt <- data.table(
  var1 = c(1, 0, NA, 0),
  var2 = c(0, 1, 1, NA)
)

# Single source variable
is_case(dt, "var1", "case1")

# Multiple source variables
is_case(dt, c("var1", "var2"), "case2")

}
